requirements.txt
python-dotenv
selenium
webdriver-manager
psycopg2-binary
pandas
openpyxl